add googlemap to use mapview
add googlemap to use mapview
